third_party/blink/web_tests/platform/generic/fast/js/pic/undictionary-expected.txt

This page tests for a caching error when transitioning prototypes away from being dictionaries. If the test passes, you'll see a series of PASS messages below.

PASS: getB(o) should be b and is.
PASS: getB(o) should be b and is.
PASS: getB(o) should be b and is.
PASS: getB(o2) should be b and is.
PASS: getB(o2) should be b and is.
PASS: getB(o2) should be b and is.
PASS: getB(o) should be b and is.
PASS: getB(o) should be b and is.
PASS: getB(o) should be b and is.
PASS: getB(o2) should be b and is.
PASS: getB(o2) should be b and is.
PASS: getB(o2) should be b and is.
